🔒 chore(pyproject.toml): update Flask and Python version requirements (#33)
* 🔒 chore(pyproject.toml): update Flask and Python version requirements
🐛 fix(request_test.py): update URL assertion to match new URL format
The Flask and Python version requirements have been updated to ensure compatibility with the latest versions of Flask and Python. The URL assertion in the request_test.py file has been updated to match the new URL format, which no longer includes the "http://localhost" prefix.

* drop python 3.6, 3.7 versions

* generate lock

* 🔨 chore(test.yml): upgrade setup-poetry action to v8
The setup-poetry action has been upgraded to version 8 to ensure that the latest version is being used. This will help to avoid any potential issues that may arise from using an outdated version of the action.

* 🔒 chore(test.yml): update poetry version to 1.4.2
This commit updates the version of Poetry used in the test workflow from 1.1.15 to 1.4.2. This is done to ensure that the latest version of Poetry is used, which may contain bug fixes and security updates.

* update black