feat(mvc_flask): add InputMethodHelper class to handle HTML-related operations
The InputMethodHelper class is added to the mvc_flask/helpers/html module. This class provides methods for generating HTML input elements with specific methods (like PUT and DELETE) that are not natively supported by HTML forms. The class includes the following methods:

- input_hidden_method: Determines the appropriate HTML string to return based on the given method string.
- _input_html: Generates a hidden HTML input element.
- _put: Generates a hidden input field for the PUT method.
- _delete: Generates a hidden input field for the DELETE method.

This class is intended to be used in the FlaskMVC class in the mvc_flask/__init__.py file. The inject_stage_and_region method in the FlaskMVC class now uses the InputMethodHelper class to generate the appropriate HTML for the method attribute in the returned dictionary.